chore: move validateService in Did.chain
