+++ packages/did/src/Did.utils.ts
@@ -9,17 +9,16 @@ import { blake2AsU8a, encodeAddress } from '@polkadot/util-crypto'
 
 import {
   DidResourceUri,
-  DidServiceEndpoint,
   DidUri,
   DidVerificationKey,
   EncryptionAlgorithms,
   EncryptionKeyType,
+  KiltAddress,
   SigningAlgorithms,
   UriFragment,
   VerificationKeyType,
-  KiltAddress,
 } from '@kiltprotocol/types'
-import { SDKErrors, ss58Format, DataUtils } from '@kiltprotocol/utils'
+import { DataUtils, SDKErrors, ss58Format } from '@kiltprotocol/utils'
 
 /// The latest version for KILT light DIDs.
 export const LIGHT_DID_LATEST_VERSION = 1
@@ -222,37 +221,6 @@ export function isKiltDidUri(
 }
 
 /**
- * Checks if a string is a valid URI according to RFC#3986.
- *
- * @param str String to be checked.
- * @returns Whether `str` is a valid URI.
- */
-export function isUri(str: string): boolean {
-  try {
-    const url = new URL(str) // this actually accepts any URI but throws if it can't be parsed
-    return url.href === str || encodeURI(decodeURI(str)) === str // make sure our URI has not been converted implicitly by URL
-  } catch {
-    return false
-  }
-}
-
-const UriFragmentRegex = /^[a-zA-Z0-9._~%+,;=*()'&$!@:/?-]+$/
-
-/**
- * Checks if a string is a valid URI fragment according to RFC#3986.
- *
- * @param str String to be checked.
- * @returns Whether `str` is a valid URI fragment.
- */
-export function isUriFragment(str: string): boolean {
-  try {
-    return UriFragmentRegex.test(str) && !!decodeURIComponent(str)
-  } catch {
-    return false
-  }
-}
-
-/**
  * Remove the `#` prefix from the UriFragment string, typically an ID.
  *
  * @param id The input ID to strip.
@@ -260,34 +228,6 @@ export function isUriFragment(str: string): boolean {
  */
 export function stripFragment(id: UriFragment): string {
   return id.replace(/^#/, '')
-}
-
-/**
- * Performs sanity checks on service endpoint data, making sure that the following conditions are met:
- *   - The `id` property is a string containing a valid URI fragment according to RFC#3986, not a complete DID URI.
- *   - If the `uris` property contains one or more strings, they must be valid URIs according to RFC#3986.
- *
- * @param endpoint A service endpoint object to check.
- */
-export function validateService(endpoint: DidServiceEndpoint): void {
-  const { id, serviceEndpoint } = endpoint
-  if (id.startsWith('did:kilt')) {
-    throw new SDKErrors.DidError(
-      `This function requires only the URI fragment part (following '#') of the service ID, not the full DID URI, which is violated by id "${id}"`
-    )
-  }
-  if (!isUriFragment(stripFragment(id))) {
-    throw new SDKErrors.DidError(
-      `The service ID must be valid as a URI fragment according to RFC#3986, which "${id}" is not. Make sure not to use disallowed characters (e.g. whitespace) or consider URL-encoding the desired id.`
-    )
-  }
-  serviceEndpoint.forEach((uri) => {
-    if (!isUri(uri)) {
-      throw new SDKErrors.DidError(
-        `A service URI must be a URI according to RFC#3986, which "${uri}" (service id "${id}") is not. Make sure not to use disallowed characters (e.g. whitespace) or consider URL-encoding resource locators beforehand.`
-      )
-    }
-  })
 }
 
 export function getAddressByKey({
